chore: add clipboard render controller
